Add files via upload
This is the rough-draft model for the home page of our assignment.  The "Home" page will include a search drop-down box for selecting type of search, a text box for user input search, a slide show bar to include graphics, promos and other information to catch the user's eye, Side columns containing video and graphics and the "About" and "Contact" links in the footer.